Compiler(Table): fix bug with resizing and arrayMap().
Previously, each time an item was added to the table, a pointer to it
was pushed to an array.
The problem with that is that when the table is resized, the location of
each item changes, but the array wasn't updated.
A much better solution is to remove the array and simply iterate over
the bucket array skipping over empty buckets.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>