Compiler(Scanner+Parser+Token): don't save all tokens.
Previously the Scanner lexed all the files at once and returned all of
their tokens in an Array. This also meant that tokens had to be heap
allocated because of how the dynamic array works.
Now the tokens aren't heap allocated anymore and the Scanner scans one
token at a time.
This also required to change how the parser kept track of the previous
token.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>