AST: Add modules.
ASTProgram now has an array for modules, and each modules has an array
of objects (functions, global variables & records).
A module is referenced with a ModuleID (an index into the modules array
in ASTProgram) and a pointer to the actual module can be requested with
the astProgramGetModule(program, module_id) function.

The parser currently creates the root modules (named ___root___) and
puts all the parsed functions there.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>