Array: Add arrayMapIndex() and use it.
Previously arrayMap() recieved a callback that
had the index of the element passed to it.
This is only needed in very specific situations,
and makes all the rest of the uses cluttered with UNUSED().

This commit reverts arrayMap() to only pass the item and user data
to the callback and adds arrayMapIndex() for the cases
where the index of the item is required.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>