Types+Parser+Validator: remove type size
The size of a type isn't needed by the
frontend (Scanner, Parser, & Validator),
only by some backends (Code generator).
The current code generator generates C, and so
it uses C types. because of that it doesn't care
about type sizes.
When a code generator that does care about
sizes (such as an assembly code generator) is added,
it will have to calculate type sizes by itself.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>