Validator: use parsed scope to look up objects.
Since objects are validated in place (meaning a new checked object
is NOT created, the same one is re-used), it is ok to use the parsed
objects when getting an object visible in the current scope for example.
Even if the object returned isn't validated yet, once the validator pass
is done, that memory address will point to a validated object.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>