Validator+tests: fix var assigned to itself at decl test.
The specific check for this case with the custom error message
would never trigger since the "id doesn't exist" error when
the initializer was validated would trigger first.
I think that this error is good enough, so I simply removed
the specific test and updated the test's expected error message.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>