Parser: only allow defer in function scope.
Since runtime defers will be added later.
For now, defers are "static", meaning the idea
is to generate them at the function epilogue, before
the return.
Allowing defers inside loops (for example), would require
runtime code to add the defer to a stack, and then in the
function epilogue the defers are popped and executed.
I have written a possible implementation for this runtime
(see files/defer_new.c), however I don't want to add anything
as complex before the code generator is written and working
with simple "static" defers.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>