Validator: validateType(): don't crash on ecursive struct.
Recursive structs will be detected in the Typechecker.
But even if they were detected here, validateType()
would still crash if a struct contains a pointer
to a struct that contains it.

The crash was due to a stack overflow caused
by infinite recursion when validating the struct field
types.
For example:
================
struct T {v:V;}
struct V{t:T;}
================
T would be validated, then V (b/c of the field) then T
again (b/c of the field), then V then T and so on.

Signed-off-by: Itai-Nelken <itainelken@gmail.com>